:sparkles: Feat: (3-1 완료) 로그인 로직 변경

+++ ProjectManager/ProjectManager/Database/Entities/Task/TaskDTO.swift
@@ -13,5 +13,5 @@ struct TaskDTO: Codable {
     var title: String
     var content: String
     var date: Date
-    var state: Int8
+    var state: Int
 }
